upload/requirements.txt
Flask==2.0.1
google-cloud-storage==2.9.0
Werkzeug==2.1.2